Add pyright to pre-commit, update readme (#106)
I added pyright to the pre-commit, so that we are warned when we create
a commit (as opposed to after pushing and when CI runs) when there is a
type error.

Unfortunately, the [pyright
pre-commit](https://github.com/RobertCraigie/pyright-python#pre-commit)
is installed in its own virtualenv and doesn't
[support](https://github.com/pre-commit/pre-commit/issues/730) using
`requirement.txt`, so it would have been cumbersome to duplicate all our
dependencies in the pre-commit configuration file. So I've instead
modified the CI to also use a virtual env, and modified the pyright
configuration to use that virtual env.

The upside of this is that the CI setup steps is identical to the local
development setup steps. So we can point new developers to the CI file
and say "just copy those steps".