Thread a Config class through all transforms (#114)
This is a step towards #11 : this PR threads a new object of type
`Config` through all transforms. Currently the information read from
`times_mappings.txt` and `times-info.json` are kept in the `Config`
class. In future we can also add things like `VedaProcessSets` to this
class, and any other configuration option / meta information that we
need to carry through the pipeline of transforms.