CI: install tool editable to fix regression testing (#117)
#110 actually broke our regression tests in CI -- since it installed the
tool non-editable, it was using the same version of the tool on the PR
branch and on main! This PR should fix it, and I hope we haven't made
any regressions in the meantime. :) [edit: I checked, no regressions!]


I tested it locally by introducting a regression and running
`run_benchmarks.py` the same way we do on CI.