Optimize transforms on dicts (#131)
Many of the transforms that have the following signature
```python
def foo(
    config: datatypes.Config, tables: Dict[str, DataFrame]
) -> Dict[str, DataFrame]:
```
are unnecessarily creating a new dictionary `output` and returning it.
Python dictionaries are mutable:
```python
In [1]: d = {1: 2}

In [2]: def foo(dic):
   ...:     dic[1] = 3
   ...:     return dic
   ...: 

In [3]: d1 = foo(d)

In [4]: d
Out[4]: {1: 3}

In [5]: d1
Out[5]: {1: 3}

In [6]: d is d1
Out[6]: True
```
so it is sufficient to modify `tables` directly.

Also, some transforms iterate through all the items in the input
dictionary just to find a single/a few key(s). We can instead do a
dictionary lookup which is a faster constant-time operation.