Feature/benchmarks on windows (#184)
This PR contains changes necessary to run the benchmarks locally on a
Windows machine.

Main changes:

- Introduced `shell=True` param to subprocess calls. Without this, the
virtualenv was not accessible to the subprocesses, which would fail
with, e.g. `ModuleNotFound: pandas` exceptions.

- Python subprocesses in windows are heavyweight and so use a lot more
RAM than linux, so I've added a variable in `utils` that limits the
number `max_workers` param in ProcessPoolExecutor constructors to a
small number (used to be the number of cpus). Without this limitation,
the pool-inside-a-pool structure of the benchmarks was spawning 32*32
python.exe processes and using all my 64Gb of RAM in a few seconds.

- Added a `--debug` switch to `run_benchmarks` to run each benchmakr as
a function call, allowing use of debuggers/breakpoints

- CI now runs CSV-only regression tests if GAMS license isn't in secrets