CI: bring back subprocess to fix regression tests (#197)
Now I remember why I used subprocess to call `xl2times` from
`run_benchmarks.py`: in the CI, when it switches to the main branch, it
needs to run the main branch's version of the tool. But if we call the
tool as a python function, I think some of the PR version of the tool
remains in memory, and we don't get what we want. This looks to be the
reason CI is failing on this PR #183 :

https://github.com/etsap-TIMES/xl2times/actions/runs/8020431584/job/21910275023

(The error is that it can't find a file in `xl2times/config/...` that
was added by the PR when it is running tests in the main branch.)

This PR undoes the change from #193 that removed the use of subprocess.
See also
https://github.com/etsap-TIMES/xl2times/pull/193/files?diff=unified&w=0#r1498683705