Cache extracted EmbeddedXlTables based on xlsx file hashes (#196)
This PR adds a feature to cache the extracted EmbeddedXlTables from
input XLSX files by default (and removes the `--use_pkl` flag). This is
because on some bigger benchmarks `openpyxl` takes ages to extract the
tables, and the xl2times developer workflow rarely involves changing the
XLSX files.

The cache is stored in `xl2times/.cache/` and saves files based on the
hash of the XLSX file. It also checks filename (and can be extended to
check modification time) to avoid hash collisions. Another future
improvement can be to have a maximum size for the cache directory.

~Change in runtime appears to be about 50% on CI (note that for this
PR's CI, the cache is created when the tool is run first on the PR
branch, and then the cache is used when it is run on `main`):~
(See comment below for more accurate performance implications)
<img width="818" alt="image"
src="https://github.com/etsap-TIMES/xl2times/assets/10712637/aa600b9a-4faf-4071-833f-4aba681492fe">


There's also a new flat `--no_cache` to ignore the cache directory and
read the files from scratch.

@SamRWest just tagging you so you are aware of this (potentially
confusing) behaviour change.

---------

Co-authored-by: Olexandr Balyk <ob@facilitate.energy>