Set output float precision & relative epsilon for DD comparison (#252)
An attempted fix for #246, which was first noticed in #240. With some
[trial and
error](https://github.com/etsap-TIMES/xl2times/pull/240#issuecomment-2507218132)
I set a precision of `.10g` (10 significant figures in
exponential/scientific notation) for floating point numbers.

I also had to change the `gdxdiff` options to additionally use a
tolerance in relative difference when comparing the output DD files to
the ground truth, otherwise the higher precision output above increased
the DD diff. We now use `1e-6` for both absolute `Eps` and relative
`RelEps` difference tolerance. If I understand correctly how `gdxdiff`
[works](https://www.gams.com/latest/docs/T_GDXDIFF.html#GDXDIFF_OPTIONS_CRITERIONEXPLANATION),
this feels like a reasonable tolerance:
```
AbsDiff := Abs(V1 - V2);
if  AbsDiff <= EpsAbsolute
then
  Result := true
else
  if EpsRelative > 0.0
  then
     Result := AbsDiff / (1.0 + DMin(Abs(V1), Abs(V2))) <= EpsRelative
  else
     Result := false;
```

Finally, we were reporting the number of lines in the output of
`gdxdiff`, but that's only the number of sets/parameters that are
different, and doesn't track improvements in number of rows different.
I'm now passing the output of `gdxdiff` to `gdxdump` so that the "GDX
Diff" column is more accurate and tracks number of rows in the diff.