Update notebook demo to run on Colab; add explanatory text (#335)
This PR updates the notebook demo to work on Google Colab. It also adds
some text to explain what the demo is about and what else can be done
(but just a sketch, please suggest how to extend/improve it!). The Demo
3 model is downloaded from a link to my drive.

Here is a shareable link to open the notebook from this branch on Colab:

https://colab.research.google.com/github/etsap-TIMES/xl2times/blob/sid/colab-demo/tests/notebook-demo.ipynb